4-deploy-to-snowflake/09-data-viewer-native-app/requirements.txt
pandas
streamlit
plotly
snowflake-snowpark-python